+++ zcncore/transaction_mobile.go
@@ -4,12 +4,18 @@
 package zcncore
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
+	"net/http"
 	"strconv"
 	"time"
 
 	"github.com/0chain/errors"
+	"github.com/0chain/gosdk/core/block"
+	"github.com/0chain/gosdk/core/encryption"
+	"github.com/0chain/gosdk/core/transaction"
+	"github.com/0chain/gosdk/core/util"
 )
 
 type TransactionCommon interface {
@@ -37,6 +43,62 @@ type TransactionCommon interface {
 	UpdateAllocation(allocID string, sizeDiff int64, expirationDiff int64, lock, fee string) error
 	WritePoolLock(allocID string, blobberID string, duration int64, lock, fee string) error
 	WritePoolUnlock(poolID string, fee string) error
+}
+
+// PriceRange represents a price range allowed by user to filter blobbers.
+type PriceRange struct {
+	Min int64 `json:"min"`
+	Max int64 `json:"max"`
+}
+
+// CreateAllocationRequest is information to create allocation.
+type CreateAllocationRequest struct {
+	DataShards      int        `json:"data_shards"`
+	ParityShards    int        `json:"parity_shards"`
+	Size            int64      `json:"size"`
+	Expiration      int64      `json:"expiration_date"`
+	Owner           string     `json:"owner_id"`
+	OwnerPublicKey  string     `json:"owner_public_key"`
+	Blobbers        []string   `json:"blobbers"`
+	ReadPriceRange  PriceRange `json:"read_price_range"`
+	WritePriceRange PriceRange `json:"write_price_range"`
+}
+
+type StakePoolSettings struct {
+	DelegateWallet string  `json:"delegate_wallet"`
+	MinStake       int64   `json:"min_stake"`
+	MaxStake       int64   `json:"max_stake"`
+	NumDelegates   int     `json:"num_delegates"`
+	ServiceCharge  float64 `json:"service_charge"`
+}
+
+type Terms struct {
+	ReadPrice        int64   `json:"read_price"`  // tokens / GB
+	WritePrice       int64   `json:"write_price"` // tokens / GB
+	MinLockDemand    float64 `json:"min_lock_demand"`
+	MaxOfferDuration int64   `json:"max_offer_duration"`
+}
+
+type Blobber struct {
+	ID                string            `json:"id"`
+	BaseURL           string            `json:"url"`
+	Terms             Terms             `json:"terms"`
+	Capacity          int64             `json:"capacity"`
+	Allocated         int64             `json:"allocated"`
+	LastHealthCheck   int64             `json:"last_health_check"`
+	StakePoolSettings StakePoolSettings `json:"stake_pool_settings"`
+}
+
+type AuthorizerStakePoolSettings struct {
+	DelegateWallet string  `json:"delegate_wallet"`
+	MinStake       int64   `json:"min_stake"`
+	MaxStake       int64   `json:"max_stake"`
+	NumDelegates   int     `json:"num_delegates"`
+	ServiceCharge  float64 `json:"service_charge"`
+}
+
+type AuthorizerConfig struct {
+	Fee int64 `json:"fee"`
 }
 
 func parseCoinStr(vs string) (uint64, error) {
@@ -508,4 +570,297 @@ func (t *Transaction) WritePoolUnlock(poolID string, fee string) error {
 // ConvertToValue converts ZCN tokens to value
 func ConvertToValue(token float64) string {
 	return strconv.FormatUint(uint64(token*float64(TOKEN_UNIT)), 10)
+}
+
+type ReqTimeout struct {
+	Milliseconds int64
+}
+
+func GetLatestFinalized(numSharders int, tm *ReqTimeout) (b *block.Header, err error) {
+	var result = make(chan *util.GetResponse, numSharders)
+	defer close(result)
+
+	var (
+		ctx    context.Context
+		cancel func()
+	)
+
+	if tm != nil && tm.Milliseconds > 0 {
+		ctx, cancel = context.WithTimeout(context.Background(), time.Millisecond*tm.Milliseconds)
+	} else {
+		ctx, cancel = context.Background()
+	}
+	defer cancel()
+
+	numSharders = len(_config.chain.Sharders) // overwrite, use all
+	queryFromShardersContext(ctx, numSharders, GET_LATEST_FINALIZED, result)
+
+	var (
+		maxConsensus   int
+		roundConsensus = make(map[string]int)
+	)
+
+	for i := 0; i < numSharders; i++ {
+		var rsp = <-result
+
+		Logger.Debug(rsp.Url, rsp.Status)
+
+		if rsp.StatusCode != http.StatusOK {
+			Logger.Error(rsp.Body)
+			continue
+		}
+
+		if err = json.Unmarshal([]byte(rsp.Body), &b); err != nil {
+			Logger.Error("block parse error: ", err)
+			err = nil
+			continue
+		}
+
+		var h = encryption.FastHash([]byte(b.Hash))
+		if roundConsensus[h]++; roundConsensus[h] > maxConsensus {
+			maxConsensus = roundConsensus[h]
+		}
+	}
+
+	if maxConsensus == 0 {
+		return nil, errors.New("", "block info not found")
+	}
+
+	return
+}
+
+func GetLatestFinalizedMagicBlock(numSharders int, tm *ReqTimeout) (m *block.MagicBlock, err error) {
+	var result = make(chan *util.GetResponse, numSharders)
+	defer close(result)
+
+	numSharders = len(_config.chain.Sharders) // overwrite, use all
+	queryFromShardersContext(ctx, numSharders, GET_LATEST_FINALIZED_MAGIC_BLOCK, result)
+
+	var (
+		ctx    context.Context
+		cancel func()
+	)
+
+	if tm != nil && tm.Milliseconds > 0 {
+		ctx, cancel = context.WithTimeout(context.Background(), time.Millisecond*tm.Milliseconds)
+	} else {
+		ctx, cancel = context.Background()
+	}
+	defer cancel()
+
+	var (
+		maxConsensus   int
+		roundConsensus = make(map[string]int)
+	)
+
+	type respObj struct {
+		MagicBlock *block.MagicBlock `json:"magic_block"`
+	}
+
+	for i := 0; i < numSharders; i++ {
+		var rsp = <-result
+
+		Logger.Debug(rsp.Url, rsp.Status)
+
+		if rsp.StatusCode != http.StatusOK {
+			Logger.Error(rsp.Body)
+			continue
+		}
+
+		var respo respObj
+		if err = json.Unmarshal([]byte(rsp.Body), &respo); err != nil {
+			Logger.Error(" magic block parse error: ", err)
+			err = nil
+			continue
+		}
+
+		m = respo.MagicBlock
+		var h = encryption.FastHash([]byte(respo.MagicBlock.Hash))
+		if roundConsensus[h]++; roundConsensus[h] > maxConsensus {
+			maxConsensus = roundConsensus[h]
+		}
+	}
+
+	if maxConsensus == 0 {
+		return nil, errors.New("", "magic block info not found")
+	}
+
+	return
+}
+
+func GetChainStats(tm *ReqTimeout) (b *block.ChainStats, err error) {
+	var result = make(chan *util.GetResponse, 1)
+	defer close(result)
+
+	var (
+		ctx    context.Context
+		cancel func()
+	)
+
+	if tm != nil && tm.Milliseconds > 0 {
+		ctx, cancel = context.WithTimeout(context.Background(), time.Millisecond*tm.Milliseconds)
+	} else {
+		ctx, cancel = context.Background()
+	}
+	defer cancel()
+
+	var numSharders = len(_config.chain.Sharders) // overwrite, use all
+	queryFromShardersContext(ctx, numSharders, GET_CHAIN_STATS, result)
+	var rsp *util.GetResponse
+	for i := 0; i < numSharders; i++ {
+		var x = <-result
+		if x.StatusCode != http.StatusOK {
+			continue
+		}
+		rsp = x
+	}
+
+	if rsp == nil {
+		return nil, errors.New("http_request_failed", "Request failed with status not 200")
+	}
+
+	if err = json.Unmarshal([]byte(rsp.Body), &b); err != nil {
+		return nil, err
+	}
+	return
+}
+
+func GetBlockByRound(numSharders int, round int64, tm *ReqTimeout) (b *block.Block, err error) {
+	var result = make(chan *util.GetResponse, numSharders)
+	defer close(result)
+
+	var (
+		ctx    context.Context
+		cancel func()
+	)
+
+	if tm != nil && tm.Milliseconds > 0 {
+		ctx, cancel = context.WithTimeout(context.Background(), time.Millisecond*tm.Milliseconds)
+	} else {
+		ctx, cancel = context.Background()
+	}
+	defer cancel()
+
+	numSharders = len(_config.chain.Sharders) // overwrite, use all
+	queryFromShardersContext(ctx, numSharders,
+		fmt.Sprintf("%sround=%d&content=full,header", GET_BLOCK_INFO, round),
+		result)
+
+	var (
+		maxConsensus   int
+		roundConsensus = make(map[string]int)
+	)
+
+	type respObj struct {
+		Block  *block.Block  `json:"block"`
+		Header *block.Header `json:"header"`
+	}
+
+	for i := 0; i < numSharders; i++ {
+		var rsp = <-result
+
+		Logger.Debug(rsp.Url, rsp.Status)
+
+		if rsp.StatusCode != http.StatusOK {
+			Logger.Error(rsp.Body)
+			continue
+		}
+
+		var respo respObj
+		if err = json.Unmarshal([]byte(rsp.Body), &respo); err != nil {
+			Logger.Error("block parse error: ", err)
+			err = nil
+			continue
+		}
+
+		if respo.Block == nil {
+			Logger.Debug(rsp.Url, "no block in response:", rsp.Body)
+			continue
+		}
+
+		if respo.Header == nil {
+			Logger.Debug(rsp.Url, "no block header in response:", rsp.Body)
+			continue
+		}
+
+		if respo.Header.Hash != string(respo.Block.Hash) {
+			Logger.Debug(rsp.Url, "header and block hash mismatch:", rsp.Body)
+			continue
+		}
+
+		b = respo.Block
+		b.Header = respo.Header
+
+		var h = encryption.FastHash([]byte(b.Hash))
+		if roundConsensus[h]++; roundConsensus[h] > maxConsensus {
+			maxConsensus = roundConsensus[h]
+		}
+	}
+
+	if maxConsensus == 0 {
+		return nil, errors.New("", "round info not found")
+	}
+
+	return
+}
+
+func GetMagicBlockByNumber(numSharders int, number int64, tm *ReqTimeout) (m *block.MagicBlock, err error) {
+	var result = make(chan *util.GetResponse, numSharders)
+	defer close(result)
+
+	var (
+		ctx    context.Context
+		cancel func()
+	)
+
+	if tm != nil && tm.Milliseconds > 0 {
+		ctx, cancel = context.WithTimeout(context.Background(), time.Millisecond*tm.Milliseconds)
+	} else {
+		ctx, cancel = context.Background()
+	}
+	defer cancel()
+
+	numSharders = len(_config.chain.Sharders) // overwrite, use all
+	queryFromShardersContext(ctx, numSharders,
+		fmt.Sprintf("%smagic_block_number=%d", GET_MAGIC_BLOCK_INFO, number),
+		result)
+
+	var (
+		maxConsensus   int
+		roundConsensus = make(map[string]int)
+	)
+
+	type respObj struct {
+		MagicBlock *block.MagicBlock `json:"magic_block"`
+	}
+
+	for i := 0; i < numSharders; i++ {
+		var rsp = <-result
+
+		Logger.Debug(rsp.Url, rsp.Status)
+
+		if rsp.StatusCode != http.StatusOK {
+			Logger.Error(rsp.Body)
+			continue
+		}
+
+		var respo respObj
+		if err = json.Unmarshal([]byte(rsp.Body), &respo); err != nil {
+			Logger.Error(" magic block parse error: ", err)
+			err = nil
+			continue
+		}
+
+		m = respo.MagicBlock
+		var h = encryption.FastHash([]byte(respo.MagicBlock.Hash))
+		if roundConsensus[h]++; roundConsensus[h] > maxConsensus {
+			maxConsensus = roundConsensus[h]
+		}
+	}
+
+	if maxConsensus == 0 {
+		return nil, errors.New("", "magic block info not found")
+	}
+
+	return
 }
